e2e: check a killed app leaves no daemons
